+++ folly/test/JsonTest.cpp
@@ -400,7 +400,7 @@ TEST(Json, StripComments) {
 
 BENCHMARK(jsonSerialize, iters) {
   folly::json::serialization_opts opts;
-  for (int i = 0; i < iters; ++i) {
+  for (size_t i = 0; i < iters; ++i) {
     folly::json::serialize(
       "qwerty \xc2\x80 \xef\xbf\xbf poiuy"
       "qwerty \xc2\x80 \xef\xbf\xbf poiuy"
@@ -420,7 +420,7 @@ BENCHMARK(jsonSerializeWithNonAsciiEncoding, iters) {
   folly::json::serialization_opts opts;
   opts.encode_non_ascii = true;
 
-  for (int i = 0; i < iters; ++i) {
+  for (size_t i = 0; i < iters; ++i) {
     folly::json::serialize(
       "qwerty \xc2\x80 \xef\xbf\xbf poiuy"
       "qwerty \xc2\x80 \xef\xbf\xbf poiuy"
@@ -440,7 +440,7 @@ BENCHMARK(jsonSerializeWithUtf8Validation, iters) {
   folly::json::serialization_opts opts;
   opts.validate_utf8 = true;
 
-  for (int i = 0; i < iters; ++i) {
+  for (size_t i = 0; i < iters; ++i) {
     folly::json::serialize(
       "qwerty \xc2\x80 \xef\xbf\xbf poiuy"
       "qwerty \xc2\x80 \xef\xbf\xbf poiuy"
@@ -457,19 +457,19 @@ BENCHMARK(jsonSerializeWithUtf8Validation, iters) {
 }
 
 BENCHMARK(parseSmallStringWithUtf, iters) {
-  for (int i = 0; i < iters << 4; ++i) {
+  for (size_t i = 0; i < iters << 4; ++i) {
     parseJson("\"I \\u2665 UTF-8 thjasdhkjh blah blah blah\"");
   }
 }
 
 BENCHMARK(parseNormalString, iters) {
-  for (int i = 0; i < iters << 4; ++i) {
+  for (size_t i = 0; i < iters << 4; ++i) {
     parseJson("\"akjhfk jhkjlakjhfk jhkjlakjhfk jhkjl akjhfk\"");
   }
 }
 
 BENCHMARK(parseBigString, iters) {
-  for (int i = 0; i < iters; ++i) {
+  for (size_t i = 0; i < iters; ++i) {
     parseJson("\""
       "akjhfk jhkjlakjhfk jhkjlakjhfk jhkjl akjhfk"
       "akjhfk jhkjlakjhfk jhkjlakjhfk jhkjl akjhfk"
@@ -491,7 +491,7 @@ BENCHMARK(toJson, iters) {
     "{\"old_value\":40,\"changed\":true,\"opened\":false,\"foo\":[1,2,3,4,5,6]}"
   );
 
-  for (int i = 0; i < iters; i++) {
+  for (size_t i = 0; i < iters; i++) {
     toJson(something);
   }
 }
